refactor: scope 선택 로직을 useScopeSelection 훅으로 추출
