+++ apps/client/src/widgets/home/index.ts
@@ -6,3 +6,4 @@ export * from './model/useDeleteApiKey';
 export * from './model/useGetApiKey';
 export * from './model/useUpdateApiKey';
 export * from './model/useGetAvailableScope';
+export * from './model/useScopeSelection';
