+++ apps/client/src/widgets/home/ui/ApiKeyForm/index.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useEffect, useMemo } from 'react';
+import { useEffect } from 'react';
 
 import { zodResolver } from '@hookform/resolvers/zod';
 import { authQueryKeys } from '@repo/shared/api';
@@ -21,6 +21,7 @@ import {
   useCreateApiKey,
   useGetApiKey,
   useGetAvailableScope,
+  useScopeSelection,
   useUpdateApiKey,
 } from '@/widgets/home';
 
@@ -95,69 +96,11 @@ const ApiKeyForm = ({ initialApiKeyData, initialAvailableScope, userRole }: ApiK
         ? 'API 키 갱신하기'
         : 'API 키 생성하기';
 
-  const scopeMap = useMemo(() => {
-    const map = new Map<string, string[]>();
-
-    const categories = availableKeyScope?.data?.list || [];
-    categories.forEach((category) => {
-      category.scopes.forEach(({ scope }) => {
-        if (scope.endsWith(':*')) return;
-
-        const prefix = scope.split(':')[0];
-        map.set(prefix!, [...(map.get(prefix!) ?? []), scope]);
-      });
-    });
-
-    return map;
-  }, [availableKeyScope]);
-
-  const handleScopeToggle = (scope: string) => {
-    const currentScopes = watch('scopes');
-
-    // 전체 scope 선택
-    if (scope.endsWith(':*')) {
-      const prefix = scope.split(':')[0];
-      const relatedScopes = scopeMap.get(prefix!) ?? [];
-
-      const allSelected =
-        relatedScopes.length > 0 && relatedScopes.every((id) => currentScopes.includes(id));
-
-      const nextSelected = allSelected
-        ? currentScopes.filter((id) => !relatedScopes.includes(id))
-        : Array.from(new Set([...currentScopes, ...relatedScopes]));
-
-      setValue('scopes', nextSelected, { shouldValidate: true });
-      return;
-    }
-
-    // 일반 scope 단일 토글
-    setValue(
-      'scopes',
-      currentScopes.includes(scope)
-        ? currentScopes.filter((id) => id !== scope)
-        : [...currentScopes, scope],
-      { shouldValidate: true },
-    );
-  };
-
-  const isScopeChecked = (scope: string) => {
-    const currentScopes = watch('scopes');
-
-    // 전체 scope일 경우 하위 scope 탐색
-    if (scope.endsWith(':*')) {
-      const prefix = scope.split(':')[0];
-      const relatedScopes = scopeMap.get(prefix!) ?? [];
-
-      return relatedScopes.length > 0 && relatedScopes.every((id) => currentScopes.includes(id));
-    }
-
-    return currentScopes.includes(scope);
-  };
-
-  const getIndentation = (level: string) => {
-    if (level.includes(':*')) return 'pl-0';
-    return 'pl-6';
-  };
+  const { handleScopeToggle, isScopeChecked, getIndentation } = useScopeSelection({
+    availableKeyScope,
+    watch,
+    setValue,
+  });
 
   const onSubmit = ({ scopes, description }: ApiKeyFormType) => {
     const data = { scopes, description };
